Use PATH and RUBY_LIB environment variables to specify the location of puppet
The local_setup.sh currently needs to be modified to point to your
puppet checkout in order to find puppet.  However, if you're already
using the documentation on the docs site
http://docs.puppetlabs.com/guides/from_source.html
you can use the PATH and RUBY_LIB environment variables to specify your
puppet location and not have to change directories as part of the test.